backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.43.0
pypdf==3.17.1
httpx==0.24.1
pdfplumber==0.10.4
numpy==1.24.3
faiss-cpu==1.7.4
tiktoken==0.5.2
